requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
langchain-openai>=0.0.5
langchain-google-genai>=0.0.6
google-generativeai>=0.3.0
chromadb>=0.4.0
pypdf>=3.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
sentence-transformers>=2.2.0